refactored childProcess() to util-functions
