+++ lib/manage-da-review.js
@@ -1,8 +1,6 @@
 'use strict';
-// const util = require('util');
-// const exec = util.promisify(require('child_process').exec);
-const {spawn}  = require("child_process")
 const {SchemaCrawler} = require('./SchemaCrawler')
+const {childProcess} = require('./util-functions')
 const fs = require("fs")
 const del = require('del');
 
@@ -63,32 +61,6 @@ module.exports = class ManageDaReview {
     return Promise.resolve(runResult);
   }
 
-  async childProcess(...args) {
-    const cp = spawn(...args);
-    return new Promise((resolve, reject)=> {
-      cp.on("error", (err) => {
-        reject(err);
-      });
-
-      cp.on("close", (code) => {
-        if (code == 0) {
-          return resolve(code);
-        }
-        else {
-          reject(new Error("ChildProcess: " + args + ".\nOn close with non-zero exist code: " + code));
-        }
-      });
-      cp.on("exit", (code) => {
-        if (code == 0) {
-          return resolve(code);
-        }
-        else {
-          reject(new Error("ChildProcess: " + args + ".\nOn exit with non-zero exist code: " + code));
-        }
-      });
-   });
-  }
-
   async cloneAndPushNewReviewInfo(baseDir, daReviewGitRepoUrl, gitCredentials, changeBranch, outputfiles) {
     // check or create the working directory
     if (!fs.existsSync(baseDir)){
@@ -114,24 +86,24 @@ module.exports = class ManageDaReview {
     globalGitconfigFileStream.write(`user.email=${gitCredentials.email}\n`);
     globalGitconfigFileStream.close();
 
-    this.childProcess('git',['config', '--global', 'credential.helper', "'store --file ~/.git-credentials'"])
-      .then(() => this.childProcess('git',['credential-store', '--file', '~/.git-credentials', 'store', '<', '.gitcredentials']))
-      .then(() => this.childProcess('git',['clone', daReviewGitRepoUrl, baseDir]))
-      .then(() => this.childProcess('git',['branch', changeBranch]))
-      .then(() => this.childProcess('git',['checkout', changeBranch]))
+    childProcess('git',['config', '--global', 'credential.helper', "'store --file ~/.git-credentials'"])
+      .then(() => childProcess('git',['credential-store', '--file', '~/.git-credentials', 'store', '<', '.gitcredentials']))
+      .then(() => childProcess('git',['clone', daReviewGitRepoUrl, baseDir]))
+      .then(() => childProcess('git',['branch', changeBranch]))
+      .then(() => childProcess('git',['checkout', changeBranch]))
       .then(() => {
         return repoDir = fs.readdirSync(baseDir)[0];
       })
-      .then((repoDir) => this.childProcess('cd',[repoDir]))
+      .then((repoDir) => childProcess('cd',[repoDir]))
       .then(async () => {
         for (let outputFile of outputfiles) {
-          await this.childProcess('mv',[outputFile, '.'])
+          await childProcess('mv',[outputFile, '.'])
         }
       })
-      .then(() => this.childProcess('git',['add', '.']))
-      .then(() => this.childProcess('git',['commit', '-m', "'New Scheam detail/linting information'"]))
-      .then(() => this.childProcess('git',['pull', 'origin', changeBranch, '--rebase']))
-      .then(() => this.childProcess('git',['push', '--set-upstream', daReviewGitRepoUrl, changeBranch]))
+      .then(() => childProcess('git',['add', '.']))
+      .then(() => childProcess('git',['commit', '-m', "'New Scheam detail/linting information'"]))
+      .then(() => childProcess('git',['pull', 'origin', changeBranch, '--rebase']))
+      .then(() => childProcess('git',['push', '--set-upstream', daReviewGitRepoUrl, changeBranch]))
       .catch((err) => {
         throw new Error(err)
       })
